% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wondr.r
\name{add_param}
\alias{add_param}
\title{Add a parameter to a WONDER API query call}
\usage{
add_param(wobj, name, ...)
}
\arguments{
\item{wobj}{a WONDER call object passed in from \code{wondr()} or a \code{wondr()}
pipe chain}

\item{\dots}{name=value fields to add to a complete parameter specification}
}
\description{
Add a parameter to a WONDER API query call
}
\examples{
\dontrun{
   wondr() \%>\%
   add_param("name", "value") \%>\%
   ... \%>\%
   make_query(database_id="dbid") -> result
}
}
