% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wondr.r
\name{make_query}
\alias{make_query}
\title{Issue a WONDER API query}
\usage{
make_query(wobj, database_id)
}
\arguments{
\item{wobj}{a WONDER call object passed in from \code{wondr()} or a \code{wondr()}
pipe chain}

\item{database_id}{CDC WONDER database ID}
}
\description{
Issue a WONDER API query
}
\examples{
\dontrun{
   wondr() \%>\%
   add_param("name", "value") \%>\%
   ... \%>\%
   make_query(database_id="dbid") -> result
}
}
